+++ packages/ast-core/fixtures/basic-app/src/users/users.module.ts
@@ -1,7 +1,8 @@
-import { Module } from "@nestjs/common";
+import { MiddlewareConsumer, Module, NestModule } from "@nestjs/common";
 import { UsersController } from "./users.controller";
 import { UsersService } from "./users.service";
 import { UsersRepository } from "./users.repository";
+import { AuthGuard } from "../auth/auth.guard";
 import { MailModule } from "../mail/mail.module";
 
 @Module({
@@ -10,4 +11,8 @@ import { MailModule } from "../mail/mail.module";
   providers: [UsersService, UsersRepository],
   exports: [UsersService],
 })
-export class UsersModule {}
+export class UsersModule implements NestModule {
+  configure(consumer: MiddlewareConsumer): void {
+    consumer.apply(AuthGuard).forRoutes(UsersController);
+  }
+}